+++ kamehouse-commons-core/src/main/java/com/nicobrest/kamehouse/commons/model/systemcommand/JvncSenderSystemCommand.java
@@ -12,11 +12,13 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 /**
- * JvncSender system command to send text and mouse clicks to a VNC server.
+ * JvncSender system command to send text and mouse clicks to a VNC server. When running on docker
+ * it sends the commands through ssh and kamehouse-cmd for mouse clicks only. Text strings are sent
+ * to the host directly from the webapps running in the docker container.
  *
  * @author nbrest
  */
-public class JvncSenderSystemCommand extends SystemCommand {
+public class JvncSenderSystemCommand extends KameHouseCmdSystemCommand {
 
   private static final int VNC_PORT = 5900;
 
@@ -30,6 +32,7 @@ public class JvncSenderSystemCommand extends SystemCommand {
    */
   public JvncSenderSystemCommand(String text) {
     logCommand = false;
+    executeOnDockerHost = false;
     this.text = text;
     setOutputCommand();
   }
@@ -39,6 +42,7 @@ public class JvncSenderSystemCommand extends SystemCommand {
    */
   public JvncSenderSystemCommand(String text, int sleepTime) {
     logCommand = false;
+    executeOnDockerHost = false;
     setSleepTime(sleepTime);
     this.text = text;
     setOutputCommand();
@@ -49,9 +53,11 @@ public class JvncSenderSystemCommand extends SystemCommand {
    */
   public JvncSenderSystemCommand(int positionX, int positionY, int clickCount) {
     logCommand = false;
+    executeOnDockerHost = true;
     this.positionX = positionX;
     this.positionY = positionY;
     this.clickCount = clickCount;
+    setKameHouseCmdCommands();
     setOutputCommand();
   }
 
@@ -60,10 +66,12 @@ public class JvncSenderSystemCommand extends SystemCommand {
    */
   public JvncSenderSystemCommand(int positionX, int positionY, int clickCount, int sleepTime) {
     logCommand = false;
+    executeOnDockerHost = true;
     setSleepTime(sleepTime);
     this.positionX = positionX;
     this.positionY = positionY;
     this.clickCount = clickCount;
+    setKameHouseCmdCommands();
     setOutputCommand();
   }
 
@@ -124,5 +132,25 @@ public class JvncSenderSystemCommand extends SystemCommand {
   @Override
   public String toString() {
     return JsonUtils.toJsonString(this, super.toString());
+  }
+
+  @Override
+  protected String getKameHouseCmdArguments() {
+    String host = DockerUtils.getHostname();
+    String password = getVncServerPassword();
+    StringBuilder args = new StringBuilder("-o jvncsender -host \"");
+    args.append(host);
+    args.append("\" -port ");
+    args.append(VNC_PORT);
+    args.append(" -password \"");
+    args.append(password);
+    args.append("\" -mouseClick \"");
+    args.append(positionX);
+    args.append(",");
+    args.append(positionY);
+    args.append(",");
+    args.append(clickCount);
+    args.append("\"");
+    return args.toString();
   }
 }