+++ changelog.md
@@ -14,8 +14,8 @@
 - Added button for create all audio playlists in groot
 - Moved screen controller buttons to a separate page from server management
 - Added ESC, ENTER and ALT+TAB, WIN+TAB and right arrow key press buttons in the screen controller page
-- Updated jvncsender to version 1.03
-- Moved jvncsender commands to execute direcly from webapps rather than going through kamehouse-cmd
+- Updated jvncsender to version 1.04
+- Moved jvncsender commands to execute direcly from webapps rather than going through kamehouse-cmd except mouse clicks when running on docker
 - Replaced mouse clicks from vncdo with jvncsender. Deprecated all use of vncdo
 - Fixed broken execute on docker host commands
 - Fixed bugs